+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 **autoupdate:** GitHub predefined hashes support ([#6416](https://github.com/ScoopInstaller/Scoop/issues/6416), [#6435](https://github.com/ScoopInstaller/Scoop/issues/6435))
 
+### Features
+
+- **scoop-uninstall**: Allow access to `$bucket` in uninstall scripts ([#6379](https://github.com/ScoopInstaller/Scoop/issues/6379))
+
 ### Bug Fixes
 
 - **scoop-download|install|update:** Fallback to default downloader when aria2 fails ([#4292](https://github.com/ScoopInstaller/Scoop/issues/4292))